@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' hide Action;
 import 'package:rebloc/rebloc.dart';
 
-abstract class DispatchProvider<T> {
+abstract mixin class DispatchProvider<T> {
   BuildContext get context;
 
   void dispatchAction(Action action) => StoreProvider.of<T>(context).dispatch(action);
